+++ src/main/java/com/jadebloom/goblin_api/currency/repository/CurrencyRepository.java
@@ -16,4 +16,6 @@ public interface CurrencyRepository extends JpaRepository<CurrencyEntity, Long> 
 
 	boolean existsByIdNotAndName(Long id, String name);
 
+	void deleteAllByCreator_Id(Long creatorId);
+
 }