+++ src/main/java/com/jadebloom/goblin_api/expense/repository/ExpenseCategoryRepository.java
@@ -16,4 +16,6 @@ public interface ExpenseCategoryRepository extends JpaRepository<ExpenseCategory
 
 	boolean existsByIdNotAndName(Long id, String name);
 
+	void deleteAllByCreator_Id(Long creatorId);
+
 }