+++ src/main/java/com/jadebloom/goblin_api/expense/repository/ExpenseRepository.java
@@ -18,6 +18,8 @@ public interface ExpenseRepository extends JpaRepository<ExpenseEntity, Long> {
 
 	void deleteAllByExpenseCategory_Id(Long expenseCategoryId);
 
+	void deleteAllByCreator_Id(Long creatorId);
+
 	void deleteAllByCurrency_Id(Long currencyId);
 
 }